change session, no refresh yet
