+++ src/error.js
@@ -85,3 +85,55 @@ export function clearInlineErrors(docId) {
   // @ts-expect-error
   view.dispatch({ effects: [clearErrors.of()] });
 }
+
+function showGlobalError(message) {
+  let errorEl = document.querySelector(`#global-error`);
+  if (errorEl) {
+    errorEl.innerText = message;
+  } else {
+    document.body.insertAdjacentHTML('beforeend', `<div id="global-error">${message}</div>`);
+  }
+}
+
+function showLocalError(docId, message) {
+  const slot = document.querySelector(`#slot-${docId}`);
+  let errorEl = document.querySelector(`#slot-${docId} #error-${docId}`);
+  if (errorEl) {
+    errorEl.innerText = message;
+  } else {
+    slot?.insertAdjacentHTML('beforeend', `<div class="error" id="error-${docId}">${message}</div>`);
+  }
+}
+
+// error handling
+export function setError(message, docId) {
+  console.error(message);
+  if (!docId) {
+    showGlobalError(message);
+    return;
+  }
+  // messages will either be string or InlineErrorMessage
+  if (typeof message != 'string') {
+    displayInlineErrors(docId, message);
+    return;
+  }
+  showLocalError(docId, message);
+}
+
+export function clearLocalError(docId) {
+  document.querySelector(`#slot-${docId} #error-${docId}`)?.remove();
+  clearInlineErrors(docId);
+}
+export function clearGlobalError() {
+  document.querySelector(`#global-error`)?.remove();
+}
+
+window.addEventListener('message', (event) => {
+  if (event.origin !== window.location.origin) {
+    return;
+  }
+  if (event.data.type === 'onError') {
+    const [err, docId] = event.data.msg;
+    setError(err, docId);
+  }
+});
